refactor: change forecast structure
