+++ src/layouts/CurrentCityForecast/CurrentCityForecast.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import styles from './CurrentCityForecast.module.scss';
 
-import CurrentDate from '../../../components/CurrentDate/CurrentDate';
+import CurrentDate from '../../components/CurrentDate/CurrentDate';
 import TemperatureUniteToggle from './TemperatureUniteToggle/TemperatureUniteToggle';
 
 function CurrentCityForecast() {
