+++ src/layouts/CurrentCityForecast/CurrentCityForecast.module.scss
@@ -1,4 +1,4 @@
-@import '../../../assets/styles/constants/main.scss';
+@import '../../assets/styles/constants/main.scss';
 
 .city-forecast {
   width: 30%;
