+++ src/layouts/DetailedForecast/DailyForecast/DailyForecast.jsx
