+++ src/layouts/DetailedForecast/DetailedForecast.jsx
