+++ src/layouts/DetailedForecast/DetailedForecast.module.scss
