+++ src/layouts/DetailedForecast/FavoriteCities/FavoriteCities.jsx
